+++ gtars/src/uniwig/mod.rs
@@ -150,7 +150,7 @@ pub fn run_uniwig(matches: &ArgMatches) {
         .expect("requires integer value");
 
     let bam_scale = matches
-        .get_one::<i32>("bamscale")
+        .get_one::<f32>("bamscale")
         .expect("requires int value");
 
     let score = matches.get_one::<bool>("score").unwrap_or_else(|| &false);
@@ -205,7 +205,7 @@ pub fn uniwig_main(
     zoom: i32,
     debug: bool,
     bam_shift: bool,
-    bam_scale: i32,
+    bam_scale: f32,
 ) -> Result<(), Box<dyn Error>> {
     // Must create a Rayon thread pool in which to run our iterators
     let pool = rayon::ThreadPoolBuilder::new()
@@ -645,7 +645,7 @@ fn process_bam(
     output_type: &str,
     debug: bool,
     bam_shift: bool,
-    bam_scale: i32
+    bam_scale: f32
 ) -> Result<(), Box<dyn Error>> {
     println!("Begin bam processing workflow...");
     let fp_string = filepath.to_string();
@@ -1088,7 +1088,7 @@ fn process_bw_in_threads(
     chrom_sizes_ref_path_string: &String,
     sel: &str,
     bam_shift:bool,
-    bam_scale: i32,
+    bam_scale: f32,
 ) {
     let (reader, writer) = os_pipe::pipe().unwrap();
     let write_fd = Arc::new(Mutex::new(writer));
@@ -1190,7 +1190,7 @@ fn determine_counting_func(
     sel_clone: &str,
     write_fd: Arc<Mutex<PipeWriter>>,
     bam_shift: bool,
-    bam_scale: i32,
+    bam_scale: f32,
 ) -> Result<(), BAMRecordError> {
 
     //let bam_shift: bool = true; // This is to ensure a shifted position workflow is used when doing bams
